Module termimage::util

Module containing various utility functions.

Statics

ANSI_BG_COLOURS
"Standard" ANSI background colours, in the same order as ANSI_BG_COLOUR_ESCAPES.

ANSI_BG_COLOUR_ESCAPES
ANSI background colour escapes.

ANSI_COLOURS
"Standard" ANSI colours, in the same order as ANSI_COLOUR_ESCAPES.

ANSI_COLOUR_ESCAPES
ANSI background colour escapes.

ANSI_RESET_ATTRIBUTES
Reset ANSI attributes

BMP_MAGIC
Magic number used for determining whether an image is a BMP.

GIF_MAGIC
Magic number used for determining whether an image is a GIF.

ICO_MAGIC
Magic number used for determining whether an image is an ICO.

JPEG_MAGIC
Magic number used for determining whether an image is a JPEG.

PNG_MAGIC
Magic number used for determining whether an image is a PNG.

Functions

closest_colour
Get the closest colour to the provided one out of the specified list of colours and retirn its index.

mul_str
Create a string consisting of n repetitions of what.